pyproject.toml
#### black ####################################################################
[tool.black]
line-length = 80
skip-magic-trailing-comma = true
target-version = ["py310"]

#### coverage #################################################################
[tool.coverage.run]
branch = true
data_file = ".coverage/data"
parallel = true

[tool.coverage.report]
exclude_lines = ["# pragma: no cover", "@overload"]
fail_under = 100.0
skip_covered = true
skip_empty = true

[tool.coverage.html]
directory = ".coverage/html"

#### hatch ####################################################################
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/ultimate_fastapi_tutorial"]

[tool.hatch.version]
path = "src/ultimate_fastapi_tutorial/__init__.py"

#### isort ####################################################################
[tool.isort]
atomic = true
case_sensitive = true
float_to_top = true
force_single_line = true
line_length = 80
lines_after_imports = 2
profile = "black"
remove_redundant_aliases = true
skip_gitignore = true

#### nitpick ##################################################################
[tool.nitpick]
style = [
  "https://raw.githubusercontent.com/dycw/nitpick/master/styles/common.toml",
  "https://raw.githubusercontent.com/dycw/nitpick/master/styles/3.10.toml",
  "https://raw.githubusercontent.com/dycw/nitpick/master/styles/pip-compile-generate-hashes.toml",
]

#### project ##################################################################
[project]
name = "ultimate-fastapi-tutorial"
dynamic = ["version"]
readme = "README.md"
requires-python = ">= 3.10"
authors = [{ name = "Derek Wan", email = "d.wan@icloud.com" }]
dependencies = [
  "alembic >= 1.8.1",
  "beartype >= 0.11.0",
  "dycw-utilities[inflection] >= 0.2.21",
  "fastapi >= 0.85.1",
  "gunicorn >= 20.1.0",
  "jinja2 >= 3.1.2",
  "pydantic[email] >= 1.10.2",
  "python-multipart >= 0.0.5",
  "sqlalchemy >= 1.4.41",
  "tenacity >= 8.1.0",
  "uvicorn >= 0.18.3",
]

[project.optional-dependencies]
dev = [
  "black",
  "bump2version",
  "flake8",
  "flake8-absolute-import",
  "flake8-annotations",
  "flake8-bandit",
  "flake8-bugbear",
  "flake8-comprehensions",
  "flake8-debugger",
  "flake8-eradicate",
  "flake8-executable",
  "flake8-no-pep420",
  "flake8-pie",
  "flake8-print",
  "flake8-simplify",
  "flake8-unused-arguments",
  "hatch",
  "isort",
  "pip-tools",
  "pytest",
  "pytest-cov",
  "pytest-xdist",
]

#### pyright ##################################################################
[tool.pyright]
include = ["src"]
pythonVersion = "3.10"
executionEnvironments = [{ root = "src" }]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportMissingTypeArgument = false
reportPrivateUsage = false
reportPrivateImportUsage = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUntypedBaseClass = false
reportImplicitStringConcatenation = "error"
reportMissingSuperCall = "error"
reportPropertyTypeMismatch = "error"
reportUnnecessaryTypeIgnoreComment = "error"
reportUnusedCallResult = "error"

#### pytest ###################################################################
[tool.pytest.ini_options]
addopts = [
  "-rsxX",
  "--color=auto",
  "--cov=ultimate_fastapi_tutorial",
  "--cov-config=pyproject.toml",
  "--cov-report=html",
  "--strict-markers",
  "--tb=native"
]
filterwarnings = ["error"]
looponfailroots = ["src"]
minversion = "7.0"
testpaths = ["src/tests"]
xfail_strict = true
